server/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nulltype==2.3.1
plaid-python==19.0.0
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1